Consume on process

+++ FineTable.Domain/Entities/EFineCollection.cs
@@ -11,15 +11,13 @@ namespace FineTable.Domain.Entities
 {
     public class EFineCollection
     {
-        
         public int Id { get; set; }
         public int MemberID { get; set; }
-        public int? Amount { get; set; }
+        public double? Amount { get; set; }
         public int? Days { get; set; }
         public MemberType MemberType { get; set; }
         public DateTime CreatedDate {get; set;}
         public DateTime? ReturnDate{get; set;}
-
         public FineStatus FineStatus { get; set; }
